+++ src/php/SpecificationBundle/Domain/Schema/FieldConfiguration.php
@@ -30,6 +30,7 @@ class FieldConfiguration
     private const SPECIAL_TYPE_CLASSES = [
         'group' => GroupFieldConfiguration::class,
         'stream' => StreamFieldConfiguration::class,
+        'dataSource' => StreamFieldConfiguration::class,
     ];
 
     private const NON_FIELDS = [